create AuthAPIControler and fix BaseAPIControler
