+++ app/Exceptions/Handler.php
@@ -31,7 +31,7 @@ class Handler extends ExceptionHandler
      *
      * This is a great spot to send exceptions to Sentry, Bugsnag, etc.
      *
-     * @param  \Exception  $exception
+     * @param  \Exception $exception
      * @return void
      */
     public function report(Exception $exception)
@@ -42,12 +42,24 @@ class Handler extends ExceptionHandler
     /**
      * Render an exception into an HTTP response.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \Exception  $exception
+     * @param  \Illuminate\Http\Request $request
+     * @param  \Exception $exception
      * @return \Illuminate\Http\Response
      */
     public function render($request, Exception $exception)
     {
+        if ($request->ajax()) {
+            if ($exception instanceof TokenInvalidException) {
+                return response()->json([
+                    'result_code' => 1000,
+                    'message' => 'token invalid'
+                ]);
+            }
+//            return response()->json([
+//                'result_code' => 2000,
+//                'message' => 'Token is required'
+//            ]);
+        }
         return parent::render($request, $exception);
     }
 }
