+++ app/Exceptions/TokenInvalidException.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: haminhhai
+ * Date: 7/25/18
+ * Time: 3:21 PM
+ */
+
+namespace App\Exceptions;
+
+
+class TokenInvalidException extends \Exception
+{
+
+}
