+++ routes/api.php
@@ -16,8 +16,8 @@ use Illuminate\Http\Request;
 Route::middleware('auth:api')->get('/user', function (Request $request) {
     return $request->user();
 });
-Route::post('auth/login', 'API\BaseAPIController@login');
-Route::post('auth/register', 'API\BaseAPIController@register');
+Route::post('auth/login', 'API\AuthAPIController@login');
+//Route::post('auth/register', 'API\AuthAPIController@register');
 Route::group(['middleware' => 'jwt.auth'], function () {
     Route::apiResources([
         'profile' => 'API\ProfileController',
